Move screen_char_draw to device specific trees in preparation for replacing atari routines.
