+++ src/apple2/screen.c
@@ -11,9 +11,20 @@
 #include <peekpoke.h>
 #include <tgi.h>
 #include <stdint.h>
+#include <stdlib.h>
 #include "../config.h"
+#include "../protocol.h"
 
 extern ConfigInfo config; 
+
+extern unsigned short scalex[];
+extern unsigned short scaley[];
+extern uint8_t font[];
+extern uint8_t fontm23[];
+extern uint16_t fontptr[];
+extern uint8_t FONT_SIZE_X;
+extern uint8_t FONT_SIZE_Y;
+extern padBool FastText; /* protocol.c */
 
 #define outb(addr,val)        (*(addr)) = (val)
 #define outw(addr,val)        (*(addr)) = (val)
@@ -67,4 +78,200 @@ void screen_beep(void)
 	{
 	}
     }
+}
+
+/**
+ * screen_char_draw(Coord, ch, count) - Output buffer from ch* of length count as PLATO characters
+ */
+void screen_char_draw(padPt* Coord, unsigned char* ch, unsigned char count)
+{
+  int16_t offset; /* due to negative offsets */
+  uint16_t x;      /* Current X and Y coordinates */
+  uint16_t y;
+  uint16_t* px;   /* Pointers to X and Y coordinates used for actual plotting */
+  uint16_t* py;
+  uint8_t i; /* current character counter */
+  uint8_t a; /* current character byte */
+  uint8_t j,k; /* loop counters */
+  int8_t b; /* current character row bit signed */
+  uint8_t width=FONT_SIZE_X;
+  uint8_t height=FONT_SIZE_Y;
+  uint16_t deltaX=1;
+  uint16_t deltaY=1;
+  uint8_t mainColor=1;
+  uint8_t altColor=0;
+  uint8_t *p;
+  uint8_t* curfont;
+  
+  switch(CurMem)
+    {
+    case M0:
+      curfont=font;
+      offset=-32;
+      break;
+    case M1:
+      curfont=font;
+      offset=64;
+      break;
+    case M2:
+      curfont=fontm23;
+      offset=-32;
+      break;
+    case M3:
+      curfont=fontm23;
+      offset=32;      
+      break;
+    }
+
+  if (CurMode==ModeRewrite)
+    {
+      altColor=0;
+    }
+  else if (CurMode==ModeInverse)
+    {
+      altColor=1;
+    }
+  
+  if (CurMode==ModeErase || CurMode==ModeInverse)
+    mainColor=0;
+  else
+    mainColor=1;
+
+  tgi_setcolor(mainColor);
+
+#ifdef __ATARI__
+  x=mul0625((Coord->x&0x1FF));
+  y=mul0375((Coord->y+14^0x1FF)&0x1FF);
+#else
+  x=scalex[(Coord->x&0x1FF)];
+  y=scaley[(Coord->y)+14&0x1FF];
+#endif
+  
+  if (FastText==padF)
+    {
+      goto chardraw_with_fries;
+    }
+
+  /* the diet chardraw routine - fast text output. */
+  
+  for (i=0;i<count;++i)
+    {
+      a=*ch;
+      ++ch;
+      a+=offset;
+      p=&curfont[fontptr[a]];
+      
+      for (j=0;j<FONT_SIZE_Y;++j)
+  	{
+  	  b=*p;
+	  
+  	  for (k=0;k<FONT_SIZE_X;++k)
+  	    {
+  	      if (b<0) /* check sign bit. */
+		{
+		  tgi_setcolor(mainColor);
+		  tgi_setpixel(x,y);
+		}
+
+	      ++x;
+  	      b<<=1;
+  	    }
+
+	  ++y;
+	  x-=width;
+	  ++p;
+  	}
+
+      x+=width;
+      y-=height;
+    }
+
+  return;
+
+ chardraw_with_fries:
+  if (Rotate)
+    {
+      deltaX=-abs(deltaX);
+      width=-abs(width);
+      px=&y;
+      py=&x;
+    }
+    else
+    {
+      px=&x;
+      py=&y;
+    }
+  
+  if (ModeBold)
+    {
+      deltaX = deltaY = 2;
+      width<<=1;
+      height<<=1;
+    }
+  
+  for (i=0;i<count;++i)
+    {
+      a=*ch;
+      ++ch;
+      a+=offset;
+      p=&curfont[fontptr[a]];
+      for (j=0;j<FONT_SIZE_Y;++j)
+  	{
+  	  b=*p;
+
+	  if (Rotate)
+	    {
+	      px=&y;
+	      py=&x;
+	    }
+	  else
+	    {
+	      px=&x;
+	      py=&y;
+	    }
+
+  	  for (k=0;k<FONT_SIZE_X;++k)
+  	    {
+  	      if (b<0) /* check sign bit. */
+		{
+		  tgi_setcolor(mainColor);
+		  if (ModeBold)
+		    {
+		      tgi_setpixel(*px+1,*py);
+		      tgi_setpixel(*px,*py+1);
+		      tgi_setpixel(*px+1,*py+1);
+		    }
+		  tgi_setpixel(*px,*py);
+		}
+	      else
+		{
+		  if (CurMode==ModeInverse || CurMode==ModeRewrite)
+		    {
+		      tgi_setcolor(altColor);
+		      if (ModeBold)
+			{
+			  tgi_setpixel(*px+1,*py);
+			  tgi_setpixel(*px,*py+1);
+			  tgi_setpixel(*px+1,*py+1);
+			}
+		      tgi_setpixel(*px,*py); 
+		    }
+		}
+
+	      x += deltaX;
+  	      b<<=1;
+  	    }
+
+	  y+=deltaY;
+	  x-=width;
+	  ++p;
+  	}
+
+      Coord->x+=width;
+      x+=width;
+      y-=height;
+    }
+
+  return;
+  
 }
